handles the setting and unsetting of debug-keys

Keys which begin with 2 - are considered boolean flags, and must be provided in the format --key=<boolean>, where for boolean the following values are allowed:

T/TRUE/1

F/FALSE/0

Usage

dev_key_handler(input, DATA, DEBUGKEYS, session, use_logical_cores)

Arguments

input

respective shiny-component

DATA

respective shiny-component

DEBUGKEYS

respective shiny-component

session

shiny session-object

use_logical_cores

internal flag

Value

modified DEBUGKEYS?

Details

Keys which begin with 3 - are expecting non-boolean, custom input, which will differ based on the respective key.