Handle saving application-state to file

assumes input$save_state to be properly formatted, as returned by shinyFiles::shinyFileSave()

assumes volumes to be the return-value of shinyFiles::getVolumes()

Usage

save_error_state(
  input,
  DATA,
  DEBUGKEYS,
  FLAGS,
  volumes,
  error,
  errordir_path = NA,
  erroneous_callback = ""
)

Arguments

input

respective shiny-component

DATA

respective shiny-component

DEBUGKEYS

respective shiny-component

FLAGS

respective shiny-component

volumes

volumes present on the machine

error

error-object

errordir_path

directory to write error-replication state to

Value

path to error-state file which may be used to load the error-triggering settings.