write results of the analysis to file

write results of the analysis to file

Usage

store_results_to_file(
  results,
  results_path,
  save_to_xlsx = FALSE,
  set_author_xlsx = FALSE
)

Arguments

results

object to write.

results_path

path to write file to, sans the filesuffix.

save_to_xlsx

should results be written into a .xlsx-file? If false, it will be written into a .csv-file instead.

set_author_xlsx

when writing to an xlsx-file, should the author be set to generated with 'duflor' via 'duflor.gui', on user_machine XX?

Value

list containing

boolean check whether or not file was written to disk successfully

complete path to the resulting file.

Note

XLSX-files are the suggested output if any additional postprocessing is required. If the data will not be modified again, writing it to csv will save space.